transmit cycle
