+++ lib/auth.js
@@ -80,7 +80,12 @@ internals.Auth.prototype.default = function (options) {
     Hoek.assert(!this.settings.default, 'Cannot set default strategy more than once');
     options = Schema.apply('auth', options, 'default strategy');
 
-    this.settings.default = this._setupRoute(Hoek.clone(options));      // Can change options
+    this.settings.default = this._setupRoute(Hoek.clone(options));      // Prevent changes to options
+
+    const routes = this.server._router.table();
+    for (let i = 0; i < routes.length; ++i) {
+        routes[i].rebuild();
+    }
 };
 
 
@@ -205,6 +210,29 @@ internals.Auth.prototype.lookup = function (route) {
 };
 
 
+internals.Auth.prototype._enabled = function (route, type) {
+
+    const config = this.lookup(route);
+    if (!config) {
+        return false;
+    }
+
+    if (type === 'authenticate') {
+        return true;
+    }
+
+    for (let i = 0; i < config.strategies.length; ++i) {
+        const name = config.strategies[i];
+        const strategy = this._strategies[name];
+        if (strategy.methods[type]) {
+            return true;
+        }
+    }
+
+    return false;
+};
+
+
 internals.Auth.authenticate = function (request) {
 
     const auth = request.server.auth;
@@ -232,10 +260,6 @@ internals.Auth.access = function (request, route) {
 internals.Auth.prototype._authenticate = function (request) {
 
     const config = this.lookup(request.route);
-    if (!config) {
-        return;
-    }
-
     return internals.authenticate(config, request, this);
 };
 
@@ -279,9 +303,7 @@ internals.Auth.payload = async function (request) {
 internals.Auth.response = async function (request) {
 
     const auth = request.server.auth;
-    const config = auth.lookup(request.route);
-    if (!config ||
-        !request.auth.isAuthenticated ||
+    if (!request.auth.isAuthenticated ||
         request.auth.strategy === 'bypass') {
 
         return;
@@ -294,7 +316,10 @@ internals.Auth.response = async function (request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    return await request.server._responder.execute(strategy.methods.response, request, { bind, realm, continue: 'undefined' });
+    const error = await request.server._responder.execute(strategy.methods.response, request, { bind, realm, continue: 'undefined' });
+    if (error) {
+        throw error;
+    }
 };
