+++ lib/cors.js
@@ -145,17 +145,10 @@ internals.handler = function (request, responder) {
 };
 
 
-exports.headers = function (response) {
-
-    const request = response.request;
-    if (request._route._special) {
-        return;
-    }
+exports.headers = function (request) {
 
     const settings = request.route.settings.cors;
-    if (!settings) {
-        return;
-    }
+    const response = request.response;
 
     response.vary('origin');
