+++ lib/request.js
@@ -95,7 +95,7 @@ internals.Request = class {
         this.setMethod = this._setMethod;
 
         this._setUrl(req.url, this.server.settings.router.stripTrailingSlash);      // Sets: this.url, this.path, this.query
-        this._setMethod(req.method);                                                    // Sets: this.method
+        this._setMethod(req.method);                                                // Sets: this.method
 
         this.id = now + ':' + server.info.id + ':' + server._requestCounter.value++;
         if (server._requestCounter.value > server._requestCounter.max) {
@@ -105,7 +105,7 @@ internals.Request = class {
         this.app = (options.app ? Hoek.shallow(options.app) : {});              // Place for application-specific state without conflicts with hapi, should not be used by plugins
         this.plugins = (options.plugins ? Hoek.shallow(options.plugins) : {});  // Place for plugins to store state without conflicts with hapi, should be namespaced using plugin name
 
-        this._route = this.server._router.specials.notFound.route;    // Used prior to routing (only settings are used, not the handler)
+        this._route = this.server._router.specials.notFound.route;              // Used prior to routing (only settings are used, not the handler)
         this.route = this._route.public;
 
         this.auth = {
