+++ lib/route.js
@@ -11,6 +11,7 @@ const Auth = require('./auth');
 const Cors = require('./cors');
 const Ext = require('./ext');
 const Handler = require('./handler');
+const Headers = require('./headers');
 const Schema = require('./schema');
 const Security = require('./security');
 const Streams = require('./streams');
@@ -208,6 +209,7 @@ internals.Route.prototype.rebuild = function (event) {
 
     if (this._special) {
         this._postCycle = (this._extensions.onPreResponse.nodes ? [this._extensions.onPreResponse] : []);
+        this._marshalCycle = [Headers.type, Headers.content];
         return;
     }
 
@@ -229,15 +231,14 @@ internals.Route.prototype.rebuild = function (event) {
         this._cycle.push(this._extensions.onPreAuth);
     }
 
-    const authenticate = (this.settings.auth !== false);                          // Anything other than 'false' can still require authentication
-    if (authenticate) {
+    if (this.server.auth._enabled(this, 'authenticate')) {
         this._cycle.push(Auth.authenticate);
     }
 
     if (this.method !== 'get') {
         this._cycle.push(internals.payload);
 
-        if (authenticate) {
+        if (this.server.auth._enabled(this, 'payload')) {
             this._cycle.push(Auth.payload);
         }
     }
@@ -286,6 +287,25 @@ internals.Route.prototype.rebuild = function (event) {
 
     if (this._extensions.onPreResponse.nodes) {
         this._postCycle.push(this._extensions.onPreResponse);
+    }
+
+    this._marshalCycle = [Headers.type];
+
+    if (this.settings.cors) {
+        this._marshalCycle.push(Cors.headers);
+    }
+
+    if (this.settings.security) {
+        this._marshalCycle.push(Security.headers);
+    }
+
+    this._marshalCycle.push(Headers.unmodified);
+    this._marshalCycle.push(Headers.cache);
+    this._marshalCycle.push(Headers.state);
+    this._marshalCycle.push(Headers.content);
+
+    if (this.server.auth._enabled(this, 'response')) {
+        this._marshalCycle.push(Auth.response);                            // Must be last in case requires access to headers
     }
 };
