+++ lib/security.js
@@ -61,30 +61,28 @@ exports.route = function (settings) {
 };
 
 
-exports.headers = function (response) {
+exports.headers = function (request) {
 
-    const request = response.request;
+    const response = request.response;
+    const security = response.request.route.settings.security;
 
-    const security = request.route.settings.security;
-    if (security) {
-        if (security._hsts) {
-            response._header('strict-transport-security', security._hsts, { override: false });
-        }
+    if (security._hsts) {
+        response._header('strict-transport-security', security._hsts, { override: false });
+    }
 
-        if (security._xframe) {
-            response._header('x-frame-options', security._xframe, { override: false });
-        }
+    if (security._xframe) {
+        response._header('x-frame-options', security._xframe, { override: false });
+    }
 
-        if (security.xss) {
-            response._header('x-xss-protection', '1; mode=block', { override: false });
-        }
+    if (security.xss) {
+        response._header('x-xss-protection', '1; mode=block', { override: false });
+    }
 
-        if (security.noOpen) {
-            response._header('x-download-options', 'noopen', { override: false });
-        }
+    if (security.noOpen) {
+        response._header('x-download-options', 'noopen', { override: false });
+    }
 
-        if (security.noSniff) {
-            response._header('x-content-type-options', 'nosniff', { override: false });
-        }
+    if (security.noSniff) {
+        response._header('x-content-type-options', 'nosniff', { override: false });
     }
 };
