+++ lib/transmit.js
@@ -3,7 +3,6 @@
 // Load modules
 
 const Http = require('http');
-const Stream = require('stream');
 
 const Ammo = require('ammo');
 const Boom = require('boom');
@@ -11,10 +10,7 @@ const Hoek = require('hoek');
 const Shot = require('shot');
 const Teamwork = require('teamwork');
 
-const Auth = require('./auth');
-const Cors = require('./cors');
 const Response = require('./response');
-const Security = require('./security');
 
 
 // Declare internals
@@ -42,62 +38,9 @@ exports.send = async function (request) {
 
 internals.marshal = async function (request) {
 
-    const response = request.response;
-
-    Cors.headers(response);
-    internals.content(response, false);
-    Security.headers(response);
-    internals.unmodified(response);
-
-    try {
-        await internals.state(response);
-    }
-    catch (err) {
-        request._log(['state', 'response', 'error'], err);
-        request._states = {};                                           // Clear broken state
-        throw err;
-    }
-
-    internals.cache(response);
-
-    if (response._isPayloadSupported() ||
-        request.method === 'head') {
-
-        await response._marshal();
-
-        if (request.jsonp &&
-            response._payload.jsonp) {
-
-            response._header('content-type', 'text/javascript' + (response.settings.charset ? '; charset=' + response.settings.charset : ''));
-            response._header('x-content-type-options', 'nosniff');
-            response._payload.jsonp(request.jsonp);
-        }
-
-        if (response._payload.size &&
-            typeof response._payload.size === 'function') {
-
-            response._header('content-length', response._payload.size(), { override: false });
-        }
-
-        if (!response._isPayloadSupported()) {
-            response._close();                              // Close unused file streams
-            response._payload = new internals.Empty();      // Set empty stream
-        }
-
-        internals.content(response, true);
-    }
-    else {
-
-        // Set empty stream
-
-        response._close();                                  // Close unused file streams
-        response._payload = new internals.Empty();
-        delete response.headers['content-length'];
-    }
-
-    const authError = await Auth.response(request);          // Must be last in case requires access to headers
-    if (authError) {
-        throw authError;
+    for (let i = 0; i < request._route._marshalCycle.length; ++i) {
+        const func = request._route._marshalCycle[i];
+        await func(request);
     }
 };
 
@@ -341,154 +284,5 @@ internals.writeHead = function (response) {
     }
     catch (err) {
         throw Boom.boomify(err);
-    }
-};
-
-
-internals.Empty = function () {
-
-    Stream.Readable.call(this);
-};
-
-Hoek.inherits(internals.Empty, Stream.Readable);
-
-
-internals.Empty.prototype._read = function (/* size */) {
-
-    this.push(null);
-};
-
-
-internals.cache = function (response) {
-
-    const request = response.request;
-
-    if (response.headers['cache-control']) {
-        return;
-    }
-
-    const policy = request.route.settings.cache &&
-        request._route._cache &&
-        (request.route.settings.cache._statuses[response.statusCode] || (response.statusCode === 304 && request.route.settings.cache._statuses['200']));
-
-    if (policy ||
-        response.settings.ttl) {
-
-        const ttl = (response.settings.ttl !== null ? response.settings.ttl : request._route._cache.ttl());
-        const privacy = (request.auth.isAuthenticated || response.headers['set-cookie'] ? 'private' : request.route.settings.cache.privacy || 'default');
-        response._header('cache-control', 'max-age=' + Math.floor(ttl / 1000) + ', must-revalidate' + (privacy !== 'default' ? ', ' + privacy : ''));
-    }
-    else if (request.route.settings.cache) {
-        response._header('cache-control', request.route.settings.cache.otherwise);
-    }
-};
-
-
-internals.content = function (response, postMarshal) {
-
-    let type = response.headers['content-type'];
-    if (!type) {
-        if (response._contentType) {
-            const charset = (response.settings.charset && response._contentType !== 'application/octet-stream' ? '; charset=' + response.settings.charset : '');
-            response.type(response._contentType + charset);
-        }
-    }
-    else {
-        type = type.trim();
-        if ((!response._contentType || !postMarshal) &&
-            response.settings.charset &&
-            type.match(/^(?:text\/)|(?:application\/(?:json)|(?:javascript))/)) {
-
-            if (!type.match(/; *charset=/)) {
-                const semi = (type[type.length - 1] === ';');
-                response.type(type + (semi ? ' ' : '; ') + 'charset=' + (response.settings.charset));
-            }
-        }
-    }
-};
-
-
-internals.state = async function (response) {
-
-    const request = response.request;
-
-    const names = {};
-    const states = [];
-
-    const requestStates = Object.keys(request._states);
-    for (let i = 0; i < requestStates.length; ++i) {
-        const stateName = requestStates[i];
-        names[stateName] = true;
-        states.push(request._states[stateName]);
-    }
-
-    const keys = Object.keys(request.server.states.cookies);
-    for (let i = 0; i < keys.length; ++i) {
-        const name = keys[i];
-        const autoValue = request.server.states.cookies[name].autoValue;
-        if (!autoValue || names[name]) {
-            continue;
-        }
-
-        names[name] = true;
-
-        if (typeof autoValue !== 'function') {
-            states.push({ name, value: autoValue });
-            continue;
-        }
-
-        try {
-            const value = await autoValue(request);
-            states.push({ name, value });
-        }
-        catch (err) {
-            throw Boom.boomify(err);
-        }
-    }
-
-    if (!states.length) {
-        return;
-    }
-
-    let header = await request.server.states.format(states);
-    const existing = response.headers['set-cookie'];
-    if (existing) {
-        header = (Array.isArray(existing) ? existing : [existing]).concat(header);
-    }
-
-    response._header('set-cookie', header);
-};
-
-
-internals.unmodified = function (response) {
-
-    const request = response.request;
-
-    // Set headers from responder.entity()
-
-    if (request._entity.etag &&
-        !response.headers.etag) {
-
-        response.etag(request._entity.etag, { vary: request._entity.vary });
-    }
-
-    if (request._entity.modified &&
-        !response.headers['last-modified']) {
-
-        response.header('last-modified', request._entity.modified);
-    }
-
-    if (response.statusCode === 304) {
-        return;
-    }
-
-    const entity = {
-        etag: response.headers.etag,
-        vary: response.settings.varyEtag,
-        modified: response.headers['last-modified']
-    };
-
-    if (Response.unmodified(request, entity)) {
-        response.code(304);
     }
 };
